+++ SpriteSheet.cpp
@@ -16,7 +16,7 @@ void Sprite2D::Draw(Tmpl8::Surface* surface, const Tmpl8::vec2& pos) //draw the 
 		{
 			Tmpl8::Pixel color = image->GetBuffer()[(x + rect.x) + (y + rect.y) * image->GetPitch()]; //get the color of the pixel
 			if (color >> 24 > 0) // shifting 24 bits to the right to get to the alfa channel 
-				surface->Plot(x + pos.x, y + pos.y, image->GetBuffer()[(x + rect.x) + (y + rect.y) * image->GetPitch()]); //draw the sprite
+				surface->Plot(x + static_cast<int>(pos.x), y + static_cast<int>(pos.y), image->GetBuffer()[(x + rect.x) + (y + rect.y) * image->GetPitch()]); //draw the sprite
 		}
 	}
 }
